实现了0速度的角度闭环基本功能，并不丝滑。下一步需要把代码原理梳理一下，并调好参数。

+++ motor/motor.h
@@ -8,6 +8,6 @@
 
 void Motor_On(void);
 void Motor_Off(void);
-void Set_Duty(uint8_t side, int8_t duty);
+void Set_Duty(uint8_t side, float duty);
 
 #endif
